% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbsd.R
\name{kbsd}
\alias{kbsd}
\title{Kernel-based sparsity diagnostic}
\usage{
kbsd(
  data,
  int_data_list,
  disthalf_vec,
  type = "Rfast",
  kernel = "gaussian",
  parallel = TRUE
)
}
\arguments{
\item{data}{Observed data.frame}

\item{int_data_list}{List of intervened on datasets}

\item{disthalf_vec}{Named numeric vector of half-distances}

\item{type}{Either "Rfast" (fast) or "base"}

\item{kernel}{Kernel function or "gaussian"}

\item{parallel}{Logical}
}
\value{
data.frame with diagnostics
}
\description{
Kernel-based sparsity diagnostic
}
