% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbsd_plot.R
\name{kbsd_plot}
\alias{kbsd_plot}
\title{Plotting the KBSD}
\usage{
kbsd_plot(
  result_df,
  xlab = "intervention",
  ylab = "EDP",
  main = "KBSD: Adjusted Boxplots",
  ylim = c(0, max(result_df$diagnostic, na.rm = TRUE)),
  ...
)
}
\arguments{
\item{result_df}{resulting data.frame of kbsd()}

\item{...}{additional arguments passed to plotting functions}
}
\value{
plot
}
\description{
Plotting the KBSD
}
