% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_disthalk_vec.R
\name{get_disthalf_vec}
\alias{get_disthalf_vec}
\title{creates a half distance vector according to rule-of-thumb}
\usage{
get_disthalf_vec(
  data,
  interventions = NULL,
  remove = NULL,
  no_extrapolation = NULL
)
}
\arguments{
\item{data}{Observed data.frame}

\item{interventions}{a vector of strings naming all intervention variables}

\item{remove}{a vector of strings naming dimensions not to be considered, e.g. the outcome variable}

\item{no_extrapolation}{a vector of strings naming all dimensions where extrapolation is excluded, e.g. binary variables}
}
\value{
a named vector
}
\description{
creates a half distance vector according to rule-of-thumb
}
